Guards

Let actors guard your mutable variables!

With Guards you can wrap mutable variables into a :guard actor. That way they can be safely accessed from parallel threads and distributed worker processes via message passing.

Guards.Guards

Guards.version

Guards.Guard

Actors.call

Actors.cast

Guards.guard

Guards.@grd

Guards.Guards — Module

Guards

Implements a generic server Actor's protocol.

The current stable, registered version is installed with

pkg> add Guards

The development version is installed with:

pkg> add "https://github.com/JuliaActors/Guards.jl"

Guards.version — Constant

Gives the package version.

Guards.Guard — Type

Guard{T}

A parametric struct returned by guard.

Parameter/Field

T: the type of the guarded variable,

link::Union{Link,Symbol}: Link or name of guard actor.

Actors.call — Method

call(gd::Guard, f, args...)

Call a guard actor gd to execute f(var, args...) on its guarded variable var and to send respond with a deep copy of the result.

Arguments

gd::Guard: link to the :guard actor,

f: callable object taking the guarded variable var as first argument,

args...: further arguments to f.

Actors.cast — Method

cast(gd::Guard, f, args...)

Cast a message to a guard actor gd to execute f(var, args...) on its guarded variable var.

Arguments

gd::Guard: link to the :guard actor,

f: callable object taking the guarded variable var as first argument,

args...: further arguments to f.

Guards.guard — Method

guard(var;  name=nothing, pid=myid(), thrd=false, 
            sticky=false, taskref=nothing)

Start a :guard actor for the variable var and return a Guard link to it.

Parameters

var: variable to guard for,

name=nothing: if a name::Symbol is provided the server is registered and the name is returned,

pid=myid(): worker pid to create the actor on,

thrd=false: thread to create the actor on,

sticky=false: if true, the actor is created in the same thread,

taskref=nothing: if a Ref{Task} variable is provided, it gets the created Task.

Guards.@grd — Macro

@grd gd
@grd f(gd, args...)

Return a deep copy of a guarded variable var or of the result of a function call on it. This is a wrapper to call.

Parameters

gd::Guard: a link to the :guard actor,

f: callable object taking the guarded variable var as first argument,

args...: further arguments to f.